+++ frontend/src/pages/SuburbReportPage/SuburbReportPage.tsx
@@ -1,7 +1,7 @@
-import ActionButtonContainer from '@/pages/SuburbReportPage/components/ActionButtonContainer';
+import ActionButtonWrapper from '@/pages/SuburbReportPage/components/ActionButtonWrapper';
 import BannerContainer from '@/components/Banner/BannerContainer';
-import BodyContainer from '@/pages/SuburbReportPage/components/BodyContainer';
-import CardContainer from '@/pages/SuburbReportPage/components/CardConatiner';
+import BodyWrapper from '@/pages/SuburbReportPage/components/BodyWrapper';
+import CardWrapper from '@/pages/SuburbReportPage/components/CardWrapper';
 import type { AppDispatch, RootState } from '@/store';
 import { fetchSuburbReport, setSuburbId } from '@/store/slices/suburbSlice';
 import { Button, Typography } from '@mui/material';
@@ -50,31 +50,31 @@ const SuburbReportPage = () => {
           Welcome to {report.suburbName},{report.state},{report.postcode}
         </Typography>
       </BannerContainer>
-      <BodyContainer minHeight={1000}>
+      <BodyWrapper minHeight={1000}>
         {/* todo: replace with real card content */}
-        <CardContainer
+        <CardWrapper
           minHeight={300}
           title={TITLES.incomeEmployment}
-        ></CardContainer>
-        <CardContainer
+        ></CardWrapper>
+        <CardWrapper
           minHeight={300}
           title={TITLES.propertyMarketInsights}
-        ></CardContainer>
-        <CardContainer
+        ></CardWrapper>
+        <CardWrapper
           minHeight={300}
           title={TITLES.demandDevelopment}
-        ></CardContainer>
-        <CardContainer minHeight={300} title={TITLES.lifeStyle}></CardContainer>
-        <CardContainer
+        ></CardWrapper>
+        <CardWrapper minHeight={300} title={TITLES.lifeStyle}></CardWrapper>
+        <CardWrapper
           minHeight={300}
           title={TITLES.safetyScore}
-        ></CardContainer>
-        <ActionButtonContainer>
+        ></CardWrapper>
+        <ActionButtonWrapper>
           {/* todo:  replace with real action buttons */}
           <Button>save this suburb</Button>
           <Button>Export PDF</Button>
-        </ActionButtonContainer>
-      </BodyContainer>
+        </ActionButtonWrapper>
+      </BodyWrapper>
     </Layout>
   );
 };
